+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 1.3-SNAPSHOT
 ADDED:
 
 - new factory method in class AbstractApplicationFrame for set the application name
+- new factory method in class AbstractApplicationFrame for set the application configuration directory name
 
 CHANGED: